+++ components/pokedetailsmodal.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-
 import { Button } from "@nextui-org/button";
+import { Image } from "@nextui-org/image";
 import {
   Modal,
   ModalContent,
@@ -8,35 +8,86 @@ import {
   ModalBody,
   ModalFooter,
 } from "@nextui-org/modal";
+import { PokeTable } from "./poketable";
 
 export const PokeDetailsModal = ({
   isOpen,
   onOpen,
   onOpenChange,
+  setPokemon,
   selectedPokemon,
+  filteredData,
 }: {
   isOpen: boolean;
   onOpen: any;
+  setPokemon: (data: any) => void;
   onOpenChange: (isOpen: boolean) => void;
   selectedPokemon: any;
+  filteredData: any;
 }) => {
+  const onNextPokemon = (isNext: boolean) => {
+    const currIdx = filteredData
+      .map((pokeinfo: any) => pokeinfo.name)
+      .indexOf(selectedPokemon.name);
+
+    const nextIdx = isNext ? currIdx + 1 : currIdx - 1;
+
+    const nextPokemon =
+      nextIdx >= 0 && nextIdx < filteredData.length
+        ? filteredData[nextIdx].name
+        : filteredData[currIdx].name;
+
+    fetch(`https://pokeapi.co/api/v2/pokemon/${nextPokemon}`)
+      .then((res) => res.json())
+      .then((data) => {
+        setPokemon(data);
+      });
+  };
+
   return (
     <>
-      <Modal isOpen={isOpen} onOpenChange={onOpenChange}>
+      <Modal
+        backdrop="blur"
+        size="lg"
+        isOpen={isOpen}
+        onOpenChange={onOpenChange}
+      >
         <ModalContent>
           {(onClose) => (
             <>
-              <ModalHeader className="flex flex-col gap-1 capitalize">
-                {selectedPokemon.name}
+              <ModalHeader className="flex capitalize">
+                <p>
+                  {selectedPokemon.name}{" "}
+                  <span className="font-light text-md">
+                    #{selectedPokemon.id.toString().padStart(4, "0")}
+                  </span>
+                </p>
               </ModalHeader>
-              <ModalBody>
-                <p>pokedetails</p>
+              <ModalBody className="flex justify-center items-center">
+                <Image
+                  isBlurred
+                  alt="Card background"
+                  className="object-cover rounded-xl"
+                  src={`https://assets.pokemon.com/assets/cms2/img/pokedex/full/${selectedPokemon.id
+                    .toString()
+                    .padStart(3, "0")}.png`}
+                  width={130}
+                />
+                <PokeTable className="md:px-10" info={selectedPokemon} />
               </ModalBody>
-              <ModalFooter>
-                <Button color="default" variant="light" onPress={onClose}>
+              <ModalFooter className="justify-center items-center">
+                <Button
+                  color="default"
+                  variant="flat"
+                  onPress={() => onNextPokemon(false)}
+                >
                   Previous
                 </Button>
-                <Button color="secondary" onPress={onClose}>
+                <Button
+                  color="default"
+                  variant="flat"
+                  onPress={() => onNextPokemon(true)}
+                >
                   Next
                 </Button>
               </ModalFooter>