+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog Format
 NOTE: all lowercase and no ending punctuation marks
 -->
 
+- [CY] - [#5](https://github.com/cyril-deguzman/pokedex/issues/5) app(whole): modal search sort pagination
 - [CY] - [#4](https://github.com/cyril-deguzman/pokedex/issues/4) app(component): link fetched data to cards and create modal
 - [CY] - [#3](https://github.com/cyril-deguzman/pokedex/issues/3) app(component): pokecard
 - [CY] - [#2](https://github.com/cyril-deguzman/pokedex/issues/2) page(home): personalize